Make compatible with PHP 5.3 again.
The anonymous callback functions that were handed over to
preg_replace_callback called private static methods of the LinkTitles class,
which worked fine with PHP 5.5.9, but not with PHP 5.3.3.

The functions were not turned (back) into static methods, and the callback
is given in the array form, as described in the PHP documentation.